ia-model/predict-model/requirements.txt
pandas
numpy
flask
joblib
setuptools
matplotlib
scikit-learn
sklearn-pandas
lightgbm
pytz